Added prism encoding encoded.enc
Added 'prism encoding encoded.enc' and encoded copy of the encoder. 
If the decoder is run directly in python it will check and decode this file then display the decoded output.
This won't happen when the decoder is imported, only when it's run directly, feel free to remove it or change the file to anything else.